Incorporated code review comments.
(From email thread)

Line 98: "unnormalizedAuth.unpersist(false)"
This is unpersisting unnormalizedAuth before the next RDD/Graph in the lineage gets materialized, which means that you don't get the benefit of caching unnormalizedAuth.  You'll want it cached until the next item in the lineage ("authNew") gets materialized, which does not happen until the first action on that graph (line 107).  Check for this issue elsewhere too.
> Done

Test suite line 107: "val (expectedAuth, expectedHub, _) = graph.runHITS(0, n)"
This is actually only checking that (a) n is the correct number of iterations and (b) the algorithm is deterministic.  It's not quite checking what the test suite claims to check.  I'd separate this into another suite, or just remove it.
> Removed

In the test suite, I'd use smaller graphs (10 vertices or so).  Unit tests should be fast, and scaling tests can be done separately (not in this coding task).
> Done

Please follow existing style for indentation in the docs (for parameters & return types); see PageRank.scala for an example.
> Removed extra spaces from semicolon names. Code passes dev/scalastyle

`Math` is deprecated in Scala. Use `math` instead.
> Done

Use “hyperlink induced topic search” the first time you mention “HITS”, then “HITS” for the rest.
> Done

Cite the HITS paper in the doc.
> Done

HITS doc: "Unfortunately, HITS requires a global normalization operation for every update, so it is not naturally parallelizable to the vertex level. As such, [[Pregel]] cannot be used."
--> This section does not belong in a doc; I'd just remove it.
> Moved to inline code comment.

The HITS.sqerr can be made accessible in the test suite by marking it "private[lib]" rather than public.
> Done

"reduce(_ + _)" can be shortened to "sum()"
> Done

"sqerr" doc for type params is not applicable.
> Done